+++ src/transform.rs
@@ -6,65 +6,33 @@
 
 use crate::Point;
 
-use crate::floating_point::FiniteF32;
 use crate::scalar::{SCALAR_NEARLY_ZERO, Scalar};
-
-#[derive(Copy, Clone, PartialEq, Default)]
-struct TransformFlags(u8);
-
-impl TransformFlags {
-    const IDENTITY: Self    = TransformFlags(0x00);
-    const TRANSLATE: Self   = TransformFlags(0x01);
-    const SCALE: Self       = TransformFlags(0x02);
-    const SKEW: Self        = TransformFlags(0x04);
-
-    #[inline] fn has_translate(self) -> bool { self.0 & 0x01 != 0 }
-    #[inline] fn has_scale(self) -> bool { self.0 & 0x02 != 0 }
-    #[inline] fn has_skew(self) -> bool { self.0 & 0x04 != 0 }
-}
-
-impl std::ops::BitOr for TransformFlags {
-    type Output = Self;
-
-    #[inline]
-    fn bitor(self, other: Self) -> Self::Output {
-        TransformFlags(self.0 | other.0)
-    }
-}
-
-impl std::ops::BitOrAssign for TransformFlags {
-    #[inline]
-    fn bitor_assign(&mut self, other: Self) {
-        self.0 |= other.0
-    }
-}
 
 
 /// An affine transformation matrix.
 ///
 /// Stores scale, skew and transform coordinates and a type of a transform.
-///
-/// # Guarantees
-///
-/// - All values are finite.
-#[derive(Copy, Clone)]
+#[allow(missing_docs)]
+#[derive(Copy, Clone, PartialEq)]
 pub struct Transform {
-    sx: FiniteF32, kx: FiniteF32, tx: FiniteF32,
-    ky: FiniteF32, sy: FiniteF32, ty: FiniteF32,
-    flags: TransformFlags,
+    pub sx: f32,
+    pub kx: f32,
+    pub ky: f32,
+    pub sy: f32,
+    pub tx: f32,
+    pub ty: f32,
 }
 
 impl Default for Transform {
     #[inline]
     fn default() -> Self {
         Transform {
-            sx: FiniteF32::FINITE_ONE,
-            kx: FiniteF32::FINITE_ZERO,
-            ky: FiniteF32::FINITE_ZERO,
-            sy: FiniteF32::FINITE_ONE,
-            tx: FiniteF32::FINITE_ZERO,
-            ty: FiniteF32::FINITE_ZERO,
-            flags: TransformFlags::IDENTITY,
+            sx: 1.0,
+            kx: 0.0,
+            ky: 0.0,
+            sy: 1.0,
+            tx: 0.0,
+            ty: 0.0,
         }
     }
 }
@@ -79,129 +47,30 @@ impl Transform {
     /// Creates a new `Transform`.
     ///
     /// We are using column-major-column-vector matrix notation, therefore it's ky-kx, not kx-ky.
-    ///
-    /// # Checks
-    ///
-    /// - All values must be finite.
-    /// - `sx` and `sy` must not be zero.
     #[inline]
-    pub fn from_row(sx: f32, ky: f32, kx: f32, sy: f32, tx: f32, ty: f32) -> Option<Self> {
-        let sx = FiniteF32::new(sx)?;
-        let ky = FiniteF32::new(ky)?;
-        let kx = FiniteF32::new(kx)?;
-        let sy = FiniteF32::new(sy)?;
-        let tx = FiniteF32::new(tx)?;
-        let ty = FiniteF32::new(ty)?;
-        Some(Transform::from_row_safe(sx, ky, kx, sy, tx, ty))
-    }
-
-    /// Creates a new `Transform`.
-    ///
-    /// We are using column-major-column-vector matrix notation, therefore it's ky-kx, not kx-ky.
-    #[inline]
-    pub(crate) fn from_row_safe(
-        sx: FiniteF32,
-        ky: FiniteF32,
-        kx: FiniteF32,
-        sy: FiniteF32,
-        tx: FiniteF32,
-        ty: FiniteF32,
-    ) -> Self {
-        let mut m = Transform {
-            sx, kx, tx,
-            ky, sy, ty,
-            flags: TransformFlags::IDENTITY,
-        };
-        m.compute_flags();
-        m
-    }
-
-    /// Creates a new translating Transform.
-    ///
-    /// # Checks
-    ///
-    /// - All values must be finite.
-    #[inline]
-    pub fn from_translate(tx: f32, ty: f32) -> Option<Self> {
-        let tx = FiniteF32::new(tx)?;
-        let ty = FiniteF32::new(ty)?;
-        Some(Transform::from_translate_safe(tx, ty))
+    pub fn from_row(sx: f32, ky: f32, kx: f32, sy: f32, tx: f32, ty: f32) -> Self {
+        Transform { sx, ky, kx, sy, tx, ty }
     }
 
     /// Creates a new translating Transform.
     #[inline]
-    pub(crate) fn from_translate_safe(tx: FiniteF32, ty: FiniteF32) -> Self {
-        let flags = if tx != FiniteF32::FINITE_ZERO || ty != FiniteF32::FINITE_ZERO {
-            TransformFlags::TRANSLATE
-        } else {
-            TransformFlags::IDENTITY
-        };
-
-        Transform {
-            sx: FiniteF32::FINITE_ONE,  kx: FiniteF32::FINITE_ZERO, tx,
-            ky: FiniteF32::FINITE_ZERO, sy: FiniteF32::FINITE_ONE,  ty,
-            flags,
-        }
-    }
-
-    /// Creates a new scaling Transform.
-    ///
-    /// # Checks
-    ///
-    /// - All values must be finite.
-    /// - `sx` and `sy` must not be zero.
-    #[inline]
-    pub fn from_scale(sx: f32, sy: f32) -> Option<Self> {
-        let sx = FiniteF32::new(sx)?;
-        let sy = FiniteF32::new(sy)?;
-        Some(Transform::from_scale_safe(sx, sy))
+    pub fn from_translate(tx: f32, ty: f32) -> Self {
+        Transform::from_row(1.0, 0.0, 0.0, 1.0, tx, ty)
     }
 
     /// Creates a new scaling Transform.
     #[inline]
-    pub(crate) fn from_scale_safe(sx: FiniteF32, sy: FiniteF32) -> Self {
-        let flags = if sx != FiniteF32::FINITE_ONE || sy != FiniteF32::FINITE_ONE {
-            TransformFlags::SCALE
-        } else {
-            TransformFlags::IDENTITY
-        };
-
-        Transform {
-            sx,                         kx: FiniteF32::FINITE_ZERO, tx: FiniteF32::FINITE_ZERO,
-            ky: FiniteF32::FINITE_ZERO, sy,                         ty: FiniteF32::FINITE_ZERO,
-            flags,
-        }
-    }
-
-    /// Creates a new skewing Transform.
-    ///
-    /// # Checks
-    ///
-    /// - All values must be finite.
-    #[inline]
-    pub fn from_skew(kx: f32, ky: f32) -> Option<Self> {
-        let kx = FiniteF32::new(kx)?;
-        let ky = FiniteF32::new(ky)?;
-        Some(Transform::from_skew_safe(kx, ky))
+    pub fn from_scale(sx: f32, sy: f32) -> Self {
+        Transform::from_row(sx, 0.0, 0.0, sy, 0.0, 0.0)
     }
 
     /// Creates a new skewing Transform.
     #[inline]
-    pub(crate) fn from_skew_safe(kx: FiniteF32, ky: FiniteF32) -> Self {
-        let flags = if kx != FiniteF32::FINITE_ZERO || ky != FiniteF32::FINITE_ZERO {
-            TransformFlags::SKEW
-        } else {
-            TransformFlags::IDENTITY
-        };
-
-        Transform {
-            sx: FiniteF32::FINITE_ONE, kx,                        tx: FiniteF32::FINITE_ZERO,
-            ky,                        sy: FiniteF32::FINITE_ONE, ty: FiniteF32::FINITE_ZERO,
-            flags,
-        }
+    pub fn from_skew(kx: f32, ky: f32) -> Self {
+        Transform::from_row(1.0, ky, kx, 1.0, 0.0, 0.0)
     }
 
-    pub(crate) fn from_sin_cos_at(sin: f32, cos: f32, px: f32, py: f32) -> Option<Self> {
+    pub(crate) fn from_sin_cos_at(sin: f32, cos: f32, px: f32, py: f32) -> Self {
         let cos_inv = 1.0 - cos;
         Transform::from_row(
             cos, sin, -sin, cos, sdot(sin, py, cos_inv, px), sdot(-sin, px, cos_inv, py)
@@ -214,34 +83,20 @@ impl Transform {
         dst1: Point,
         dst2: Point,
     ) -> Option<Self> {
-        let tmp = from_poly2(src1, src2)?;
+        let tmp = from_poly2(src1, src2);
         let res = tmp.invert()?;
-        let tmp = from_poly2(dst1, dst2)?;
-        concat(&tmp, &res)
+        let tmp = from_poly2(dst1, dst2);
+        Some(concat(&tmp, &res))
     }
 
-    /// Returns scale pair.
     #[inline]
-    pub fn get_scale(&self) -> (f32, f32) {
-        (self.sx.get(), self.sy.get())
-    }
-
-    /// Returns skew pair.
-    #[inline]
-    pub fn get_skew(&self) -> (f32, f32) {
-        (self.kx.get(), self.ky.get())
-    }
-
-    /// Returns translate pair.
-    #[inline]
-    pub fn get_translate(&self) -> (f32, f32) {
-        (self.tx.get(), self.ty.get())
-    }
-
-    /// Returns all values.
-    #[inline]
-    pub fn get_row(&self) -> (f32, f32, f32, f32, f32, f32) {
-        (self.sx.get(), self.ky.get(), self.kx.get(), self.sy.get(), self.tx.get(), self.ty.get())
+    pub(crate) fn is_finite(&self) -> bool {
+        self.sx.is_finite() &&
+        self.ky.is_finite() &&
+        self.kx.is_finite() &&
+        self.sy.is_finite() &&
+        self.tx.is_finite() &&
+        self.ty.is_finite()
     }
 
     /// Checks that transform is identity.
@@ -249,7 +104,7 @@ impl Transform {
     /// The transform type is detected on creation, so this method is essentially free.
     #[inline]
     pub fn is_identity(&self) -> bool {
-        self.flags == TransformFlags::IDENTITY
+        *self == Transform::default()
     }
 
     /// Checks that transform is scale-only.
@@ -257,7 +112,7 @@ impl Transform {
     /// The transform type is detected on creation, so this method is essentially free.
     #[inline]
     pub fn is_scale(&self) -> bool {
-        self.flags == TransformFlags::SCALE
+        self.has_scale() && !self.has_skew() && !self.has_translate()
     }
 
     /// Checks that transform is skew-only.
@@ -265,7 +120,7 @@ impl Transform {
     /// The transform type is detected on creation, so this method is essentially free.
     #[inline]
     pub fn is_skew(&self) -> bool {
-        self.flags == TransformFlags::SKEW
+        !self.has_scale() && self.has_skew() && !self.has_translate()
     }
 
     /// Checks that transform is translate-only.
@@ -273,7 +128,7 @@ impl Transform {
     /// The transform type is detected on creation, so this method is essentially free.
     #[inline]
     pub fn is_translate(&self) -> bool {
-        self.flags == TransformFlags::TRANSLATE
+        !self.has_scale() && !self.has_skew() && self.has_translate()
     }
 
     /// Checks that transform contains only scale and translate.
@@ -281,9 +136,7 @@ impl Transform {
     /// The transform type is detected on creation, so this method is essentially free.
     #[inline]
     pub fn is_scale_translate(&self) -> bool {
-        self.flags == TransformFlags::SCALE ||
-        self.flags == TransformFlags::TRANSLATE ||
-        self.flags == TransformFlags::SCALE | TransformFlags::TRANSLATE
+        (self.has_scale() || self.has_translate()) && !self.has_skew()
     }
 
     /// Checks that transform contains a scale part.
@@ -291,7 +144,7 @@ impl Transform {
     /// The transform type is detected on creation, so this method is essentially free.
     #[inline]
     pub fn has_scale(&self) -> bool {
-        self.flags.has_scale()
+        self.sx != 1.0 || self.sy != 1.0
     }
 
     /// Checks that transform contains a skew part.
@@ -299,7 +152,7 @@ impl Transform {
     /// The transform type is detected on creation, so this method is essentially free.
     #[inline]
     pub fn has_skew(&self) -> bool {
-        self.flags.has_skew()
+        self.kx != 0.0 || self.ky != 0.0
     }
 
     /// Checks that transform contains a translate part.
@@ -307,73 +160,56 @@ impl Transform {
     /// The transform type is detected on creation, so this method is essentially free.
     #[inline]
     pub fn has_translate(&self) -> bool {
-        self.flags.has_translate()
-    }
-
-    #[inline]
-    fn compute_flags(&mut self) {
-        self.flags = TransformFlags::IDENTITY;
-
-        if self.sx != FiniteF32::FINITE_ONE || self.sy != FiniteF32::FINITE_ONE {
-            self.flags |= TransformFlags::SCALE;
-        }
-
-        if self.tx != FiniteF32::FINITE_ZERO || self.ty != FiniteF32::FINITE_ZERO {
-            self.flags |= TransformFlags::TRANSLATE;
-        }
-
-        if self.kx != FiniteF32::FINITE_ZERO || self.ky != FiniteF32::FINITE_ZERO {
-            self.flags |= TransformFlags::SKEW;
-        }
+        self.tx != 0.0 || self.ty != 0.0
     }
 
     /// Pre-scales the current transform.
     #[inline]
     #[must_use]
-    pub fn pre_scale(&self, sx: f32, sy: f32) -> Option<Self> {
-        let other = Transform::from_scale(sx, sy)?;
+    pub fn pre_scale(&self, sx: f32, sy: f32) -> Self {
+        let other = Transform::from_scale(sx, sy);
         self.pre_concat(&other)
     }
 
     /// Post-scales the current transform.
     #[inline]
     #[must_use]
-    pub fn post_scale(&mut self, sx: f32, sy: f32) -> Option<Self> {
-        let other = Transform::from_scale(sx, sy)?;
+    pub fn post_scale(&mut self, sx: f32, sy: f32) -> Self {
+        let other = Transform::from_scale(sx, sy);
         self.post_concat(&other)
     }
 
     /// Pre-translates the current transform.
     #[inline]
     #[must_use]
-    pub fn pre_translate(&self, tx: f32, ty: f32) -> Option<Self> {
-        let other = Transform::from_translate(tx, ty)?;
+    pub fn pre_translate(&self, tx: f32, ty: f32) -> Self {
+        let other = Transform::from_translate(tx, ty);
         self.pre_concat(&other)
     }
 
     /// Post-translates the current transform.
     #[inline]
     #[must_use]
-    pub fn post_translate(&self, tx: f32, ty: f32) -> Option<Self> {
-        let other = Transform::from_translate(tx, ty)?;
+    pub fn post_translate(&self, tx: f32, ty: f32) -> Self {
+        let other = Transform::from_translate(tx, ty);
         self.post_concat(&other)
     }
 
     /// Pre-concats the current transform.
     #[inline]
     #[must_use]
-    pub fn pre_concat(&self, other: &Self) -> Option<Self> {
+    pub fn pre_concat(&self, other: &Self) -> Self {
         concat(self, other)
     }
 
     /// Post-concats the current transform.
     #[inline]
     #[must_use]
-    pub fn post_concat(&self, other: &Self) -> Option<Self> {
+    pub fn post_concat(&self, other: &Self) -> Self {
         concat(other, self)
     }
 
-    pub(crate) fn from_sin_cos(sin: f32, cos: f32) -> Option<Self> {
+    pub(crate) fn from_sin_cos(sin: f32, cos: f32) -> Self {
         Transform::from_row(cos, sin, -sin, cos, 0.0, 0.0)
     }
 
@@ -384,26 +220,22 @@ impl Transform {
 
         // TODO: simd
 
-        let (tx, ty) = self.get_translate();
         if self.is_identity() {
             // Do nothing.
         } else if self.is_translate() {
             for p in points {
-                p.x += tx;
-                p.y += ty;
+                p.x += self.tx;
+                p.y += self.ty;
             }
         } else if self.is_scale_translate() {
-            let (sx, sy) = self.get_scale();
             for p in points {
-                p.x = p.x * sx + tx;
-                p.y = p.y * sy + ty;
+                p.x = p.x * self.sx + self.tx;
+                p.y = p.y * self.sy + self.ty;
             }
         } else {
-            let (sx, sy) = self.get_scale();
-            let (kx, ky) = self.get_skew();
             for p in points {
-                let x = p.x * sx + p.y * kx + tx;
-                let y = p.x * ky + p.y * sy + ty;
+                let x = p.x * self.sx + p.y * self.kx + self.tx;
+                let y = p.x * self.ky + p.y * self.sy + self.ty;
                 p.x = x;
                 p.y = y;
             }
@@ -421,21 +253,6 @@ impl Transform {
     }
 }
 
-impl std::cmp::PartialEq for Transform {
-    fn eq(&self, other: &Transform) -> bool {
-        if self.flags != other.flags {
-            false
-        } else {
-            self.sx == other.sx &&
-            self.ky == other.ky &&
-            self.kx == other.kx &&
-            self.sy == other.sy &&
-            self.tx == other.tx &&
-            self.ty == other.ty
-        }
-    }
-}
-
 impl std::fmt::Debug for Transform {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("Transform")
@@ -449,7 +266,7 @@ impl std::fmt::Debug for Transform {
     }
 }
 
-fn from_poly2(p0: Point, p1: Point) -> Option<Transform> {
+fn from_poly2(p0: Point, p1: Point) -> Transform {
     Transform::from_row(
         p1.y - p0.y,
         p0.x - p1.x,
@@ -465,24 +282,28 @@ fn invert(ts: &Transform) -> Option<Transform> {
     debug_assert!(!ts.is_identity());
 
     if ts.is_scale_translate() {
-        let (sx, _, _, sy, tx, ty) = ts.get_row();
         if ts.has_scale() {
-            let inv_x = sx.invert();
-            let inv_y = sy.invert();
-            Transform::from_row(inv_x, 0.0, 0.0, inv_y, -tx * inv_x, -ty * inv_y)
+            let inv_x = ts.sx.invert();
+            let inv_y = ts.sy.invert();
+            Some(Transform::from_row(inv_x, 0.0, 0.0, inv_y, -ts.tx * inv_x, -ts.ty * inv_y))
         } else {
             // translate only
-            Transform::from_translate(-tx, -ty)
+            Some(Transform::from_translate(-ts.tx, -ts.ty))
         }
     } else {
         let inv_det = inv_determinant(ts)?;
-        compute_inv(ts, inv_det)
+        let inv_ts = compute_inv(ts, inv_det);
+
+        if inv_ts.is_finite() {
+            Some(inv_ts)
+        } else {
+            None
+        }
     }
 }
 
 fn inv_determinant(ts: &Transform) -> Option<f64> {
-    let (sx, ky, kx, sy, _, _) = ts.get_row();
-    let det = dcross(sx as f64, sy as f64, kx as f64, ky as f64);
+    let det = dcross(ts.sx as f64, ts.sy as f64, ts.kx as f64, ts.ky as f64);
 
     // Since the determinant is on the order of the cube of the matrix members,
     // compare to the cube of the default nearly-zero constant (although an
@@ -495,26 +316,24 @@ fn inv_determinant(ts: &Transform) -> Option<f64> {
     }
 }
 
-fn compute_inv(ts: &Transform, inv_det: f64) -> Option<Transform> {
-    let (sx, ky, kx, sy, tx, ty) = ts.get_row();
-
+fn compute_inv(ts: &Transform, inv_det: f64) -> Transform {
     Transform::from_row(
-        (sy as f64 * inv_det) as f32,
-        (-ky as f64 * inv_det) as f32,
-        (-kx as f64 * inv_det) as f32,
-        (sx as f64 * inv_det) as f32,
+        (ts.sy as f64 * inv_det) as f32,
+        (-ts.ky as f64 * inv_det) as f32,
+        (-ts.kx as f64 * inv_det) as f32,
+        (ts.sx as f64 * inv_det) as f32,
         dcross_dscale(
-            kx,
-            ty,
-            sy,
-            tx,
+            ts.kx,
+            ts.ty,
+            ts.sy,
+            ts.tx,
             inv_det,
         ),
         dcross_dscale(
-            ky,
-            tx,
-            sx,
-            ty,
+            ts.ky,
+            ts.tx,
+            ts.sx,
+            ts.ty,
             inv_det,
         ),
     )
@@ -532,33 +351,29 @@ fn sdot(a: f32, b: f32, c: f32, d: f32) -> f32 {
     a * b + c * d
 }
 
-fn concat(a: &Transform, b: &Transform) -> Option<Transform> {
+fn concat(a: &Transform, b: &Transform) -> Transform {
     if a.is_identity() {
-        Some(*b)
+        *b
     } else if b.is_identity() {
-        Some(*a)
+        *a
     } else if !a.has_skew() && !b.has_skew() {
         // just scale and translate
-        let (a_sx, _, _, a_sy, a_tx, a_ty) = a.get_row();
-        let (b_sx, _, _, b_sy, b_tx, b_ty) = b.get_row();
         Transform::from_row(
-            a_sx * b_sx,
+            a.sx * b.sx,
             0.0,
             0.0,
-            a_sy * b_sy,
-            a_sx * b_tx + a_tx,
-            a_sy * b_ty + a_ty,
+            a.sy * b.sy,
+            a.sx * b.tx + a.tx,
+            a.sy * b.ty + a.ty,
         )
     } else {
-        let (a_sx, a_ky, a_kx, a_sy, a_tx, a_ty) = a.get_row();
-        let (b_sx, b_ky, b_kx, b_sy, b_tx, b_ty) = b.get_row();
         Transform::from_row(
-            mul_add_mul(a_sx, b_sx, a_kx, b_ky),
-            mul_add_mul(a_ky, b_sx, a_sy, b_ky),
-            mul_add_mul(a_sx, b_kx, a_kx, b_sy),
-            mul_add_mul(a_ky, b_kx, a_sy, b_sy),
-            mul_add_mul(a_sx, b_tx, a_kx, b_ty) + a_tx,
-            mul_add_mul(a_ky, b_tx, a_sy, b_ty) + a_ty,
+            mul_add_mul(a.sx, b.sx, a.kx, b.ky),
+            mul_add_mul(a.ky, b.sx, a.sy, b.ky),
+            mul_add_mul(a.sx, b.kx, a.kx, b.sy),
+            mul_add_mul(a.ky, b.kx, a.sy, b.sy),
+            mul_add_mul(a.sx, b.tx, a.kx, b.ty) + a.tx,
+            mul_add_mul(a.ky, b.tx, a.sy, b.ty) + a.ty,
         )
     }
 }
@@ -574,16 +389,16 @@ mod tests {
     #[test]
     fn transform() {
         assert_eq!(Transform::identity(),
-                   Transform::from_row(1.0, 0.0, 0.0, 1.0, 0.0, 0.0).unwrap());
+                   Transform::from_row(1.0, 0.0, 0.0, 1.0, 0.0, 0.0));
 
-        assert_eq!(Transform::from_scale(1.0, 2.0).unwrap(),
-                   Transform::from_row(1.0, 0.0, 0.0, 2.0, 0.0, 0.0).unwrap());
+        assert_eq!(Transform::from_scale(1.0, 2.0),
+                   Transform::from_row(1.0, 0.0, 0.0, 2.0, 0.0, 0.0));
 
-        assert_eq!(Transform::from_skew(2.0, 3.0).unwrap(),
-                   Transform::from_row(1.0, 3.0, 2.0, 1.0, 0.0, 0.0).unwrap());
+        assert_eq!(Transform::from_skew(2.0, 3.0),
+                   Transform::from_row(1.0, 3.0, 2.0, 1.0, 0.0, 0.0));
 
-        assert_eq!(Transform::from_translate(5.0, 6.0).unwrap(),
-                   Transform::from_row(1.0, 0.0, 0.0, 1.0, 5.0, 6.0).unwrap());
+        assert_eq!(Transform::from_translate(5.0, 6.0),
+                   Transform::from_row(1.0, 0.0, 0.0, 1.0, 5.0, 6.0));
 
         let ts = Transform::identity();
         assert_eq!(ts.is_identity(), true);
@@ -595,7 +410,7 @@ mod tests {
         assert_eq!(ts.has_skew(), false);
         assert_eq!(ts.has_translate(), false);
 
-        let ts = Transform::from_scale(2.0, 3.0).unwrap();
+        let ts = Transform::from_scale(2.0, 3.0);
         assert_eq!(ts.is_identity(), false);
         assert_eq!(ts.is_scale(), true);
         assert_eq!(ts.is_skew(), false);
@@ -605,7 +420,7 @@ mod tests {
         assert_eq!(ts.has_skew(), false);
         assert_eq!(ts.has_translate(), false);
 
-        let ts = Transform::from_skew(2.0, 3.0).unwrap();
+        let ts = Transform::from_skew(2.0, 3.0);
         assert_eq!(ts.is_identity(), false);
         assert_eq!(ts.is_scale(), false);
         assert_eq!(ts.is_skew(), true);
@@ -615,7 +430,7 @@ mod tests {
         assert_eq!(ts.has_skew(), true);
         assert_eq!(ts.has_translate(), false);
 
-        let ts = Transform::from_translate(2.0, 3.0).unwrap();
+        let ts = Transform::from_translate(2.0, 3.0);
         assert_eq!(ts.is_identity(), false);
         assert_eq!(ts.is_scale(), false);
         assert_eq!(ts.is_skew(), false);
@@ -625,7 +440,7 @@ mod tests {
         assert_eq!(ts.has_skew(), false);
         assert_eq!(ts.has_translate(), true);
 
-        let ts = Transform::from_row(1.0, 2.0, 3.0, 4.0, 5.0, 6.0).unwrap();
+        let ts = Transform::from_row(1.0, 2.0, 3.0, 4.0, 5.0, 6.0);
         assert_eq!(ts.is_identity(), false);
         assert_eq!(ts.is_scale(), false);
         assert_eq!(ts.is_skew(), false);
@@ -635,24 +450,24 @@ mod tests {
         assert_eq!(ts.has_skew(), true);
         assert_eq!(ts.has_translate(), true);
 
-        let ts = Transform::from_scale(1.0, 1.0).unwrap();
+        let ts = Transform::from_scale(1.0, 1.0);
         assert_eq!(ts.has_scale(), false);
 
-        let ts = Transform::from_skew(0.0, 0.0).unwrap();
+        let ts = Transform::from_skew(0.0, 0.0);
         assert_eq!(ts.has_skew(), false);
 
-        let ts = Transform::from_translate(0.0, 0.0).unwrap();
+        let ts = Transform::from_translate(0.0, 0.0);
         assert_eq!(ts.has_translate(), false);
     }
 
     #[test]
     fn concat() {
-        let mut ts = Transform::from_row(1.2, 3.4, -5.6, -7.8, 1.2, 3.4).unwrap();
-        ts = ts.pre_scale(2.0, -4.0).unwrap();
-        assert_eq!(ts, Transform::from_row(2.4, 6.8, 22.4, 31.2, 1.2, 3.4).unwrap());
+        let mut ts = Transform::from_row(1.2, 3.4, -5.6, -7.8, 1.2, 3.4);
+        ts = ts.pre_scale(2.0, -4.0);
+        assert_eq!(ts, Transform::from_row(2.4, 6.8, 22.4, 31.2, 1.2, 3.4));
 
-        let mut ts = Transform::from_row(1.2, 3.4, -5.6, -7.8, 1.2, 3.4).unwrap();
-        ts = ts.post_scale(2.0, -4.0).unwrap();
-        assert_eq!(ts, Transform::from_row(2.4, -13.6, -11.2, 31.2, 2.4, -13.6).unwrap());
+        let mut ts = Transform::from_row(1.2, 3.4, -5.6, -7.8, 1.2, 3.4);
+        ts = ts.post_scale(2.0, -4.0);
+        assert_eq!(ts, Transform::from_row(2.4, -13.6, -11.2, 31.2, 2.4, -13.6));
     }
 }