+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Reduce `Transform` strictness. It's no longer guarantee to have only finite values,
+  therefore we don't have to check each operation.
 
 ## [0.4.2] - 2021-01-23
 ### Fixed